+++ examples/approx_areas.R
@@ -1,4 +1,4 @@
-if (require(tmap) && packageVersion("tmap") >= "3.99") {
+if (require(tmap)) {
     data(NLD_muni)
 
     NLD_muni$area <- approx_areas(NLD_muni, total.area = 33893)
@@ -17,17 +17,17 @@ if (require(tmap) && packageVersion("tmap") >= "3.99") {
     }
 
     # area of the polygons
-    approx_areas(NLD_muni) %>% summary_areas()
+    summary_areas(approx_areas(NLD_muni))
 
     # area of the polygons, adjusted corrected for a specified total area size
-    approx_areas(NLD_muni, total.area=33893) %>% summary_areas()
+    summary_areas(approx_areas(NLD_muni, total.area=33893))
 
     # proportional area of the polygons
-    approx_areas(NLD_muni, target = "prop") %>% summary_areas()
+    summary_areas(approx_areas(NLD_muni, target = "prop"))
 
     # area in squared miles
-    approx_areas(NLD_muni, target = "mi mi") %>% summary_areas()
+    summary_areas(approx_areas(NLD_muni, target = "mi mi"))
 
     # area of the polygons when unprojected
-    approx_areas(NLD_muni %>% sf::st_transform(crs = 4326)) %>% summary_areas()
+    summary_areas(approx_areas(NLD_muni %>% sf::st_transform(crs = 4326)))
 }