+++ examples/palette_explorer.R
@@ -1,5 +1,0 @@
-\dontrun{
-if (require(shiny) && require(shinyjs)) {
-    palette_explorer()
-}
-}